apps/backend/requirements.txt
fastapi>=0.115.0
email-validator>=2.0.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
google-play-scraper>=1.2.7
apscheduler>=3.10.0
httpx>=0.27.0
google-generativeai>=0.8.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
